optimisation des requetes il faut revoir les test maintenant

+++ api2/app/services/user_service.py
@@ -4,7 +4,7 @@ from sqlalchemy.orm import Session
 from app.errors.user_error import UserNotFoundError, UserEmailDoesExist, UserEmailNotValide
 
 
-def insert(entity: Entity) -> str:
+def insert(entity_dict: dict) -> str:
     """
     Insère un nouvel utilisateur dans la base de données.
 
@@ -18,13 +18,14 @@ def insert(entity: Entity) -> str:
         UserEmailDoesExist : Si l'email existe déjà.
     """
     db: Session = next(get_db())
-    email: str = entity.get_email()
 
     try:
         check_email = db.query(Entity).filter(
-            Entity._email == entity.get_email()).first()
+            Entity._email == entity_dict["email"]).first()
         if check_email is not None:
             raise UserEmailDoesExist()
+        entity: Entity = Entity(entity_dict["email"])
+        entity.set_password(entity_dict["password"])
         db.add(entity)
         db.commit()
     except UserEmailDoesExist:
@@ -36,28 +37,20 @@ def insert(entity: Entity) -> str:
     finally:
         db.close()
 
-    return email
+    return entity_dict["email"]
 
 
-def delete(entity: Entity) -> str:
-    """
-    Supprime un utilisateur par email.
-
-    Paramètres:
-        entity (Entity) : L'utilisateur à supprimer.
-
-    Retourne:
-        str : L'email de l'utilisateur supprimé.
-
-    Lève:
-        UserNotFoundError : Si l'utilisateur n'existe pas.
-    """
+def delete(entity_dict: dict) -> str:
     db: Session = next(get_db())
     try:
-        entity_find = find_by_email(entity)
+        # entity_find = find_by_email(entity_dict["email"])
+        entity_find = db.query(Entity).filter(
+            Entity._email == entity_dict["email"]).first()
+        if not entity_find:
+            raise UserNotFoundError()
         db.delete(entity_find)
         db.commit()
-        return entity.get_email()
+        return entity_dict["email"]
     except UserNotFoundError:
         db.rollback()
         raise
@@ -65,67 +58,114 @@ def delete(entity: Entity) -> str:
         db.close()
 
 
-def find_by_email(entity: Entity) -> Entity:
-    """
-    Recherche un utilisateur par email.
-
-    Paramètres:
-        entity (Entity) : L'utilisateur avec l'email à rechercher.
-
-    Retourne:
-        Entity : L'utilisateur trouvé.
-
-    Lève:
-        UserNotFoundError : Si aucun utilisateur n'est trouvé.
-    """
+def find_by_email(entity_dict: dict) -> dict:
     db: Session = next(get_db())
     try:
-        result = db.query(Entity).filter(
-            Entity._email == entity.get_email()).first()
-        if not result:
+        entity = db.query(Entity).filter(
+            Entity._email == entity_dict["email"]).first()
+        if not entity:
             raise UserNotFoundError()
+        result = entity.to_dict()
+        del result["password"]
         return result
-    finally:
-        db.close()
+    except:
+        db.rollback()
+        raise
 
 
 def find_all() -> list:
-    """
-    Récupère tous les utilisateurs.
-
-    Retourne:
-        list : Liste des utilisateurs.
-    """
     db: Session = next(get_db())
     try:
-        return db.query(Entity).all()
+        entitys = db.query(Entity).all()
+        for i in range(len(entitys)):
+            entitys[i] = entitys[i].to_dict()
+            del entitys[i]["password"]
+        db.commit()
+        return entitys
+    except Exception as e:
+        raise Exception(f"An error occurred: {str(e)}")
     finally:
         db.close()
 
 
-def update(entity: Entity) -> Entity:
-    """
-    Met à jour un utilisateur.
-
-    Paramètres:
-        entity (Entity) : L'utilisateur à mettre à jour.
-
-    Retourne:
-        Entity : L'utilisateur mis à jour.
-
-    Lève:
-        UserNotFoundError : Si l'utilisateur n'existe pas.
-    """
+def replace(entity_dict: dict, email: str) -> dict:
     db: Session = next(get_db())
-    if entity.get_password() is None:
-        entity.set_password(find_by_email(entity).get_password()
-                            )  # Garde l'ancien mot de passe si non spécifié
     try:
-        if find_by_email(entity) == None:
+        entity_to_update: Entity = db.query(Entity).filter(
+            Entity._email == email).first()
+        if entity_to_update is None:
             raise UserNotFoundError()
-        db.merge(entity)
+
+        # 🔹 Mise à jour des champs si fournis
+
+        entity_to_update.set_email(entity_dict["email"])
+
+        entity_to_update.set_role(entity_dict["role"])
+
+        entity_to_update.set_password(entity_dict["password"])
+
+        entity_to_update.set_lastname(entity_dict["lastname"])
+
+        entity_to_update.set_firstname(entity_dict["firstname"])
+
+        entity_to_update.set_birth_at(entity_dict["birth_at"])
+
+        # ✅ Appliquer les modifications
+        entity_dict: dict = entity_to_update.to_dict()
+        db.add(entity_to_update)
         db.commit()
-        return find_by_email(entity)
+        db.refresh(entity_to_update
+                   )  # Rafraîchir pour récupérer les nouvelles valeurs
+        # entity_dict: dict = entity_to_update.to_dict()
+        return entity_dict
+    except UserNotFoundError:
+        raise
+    finally:
+        db.close()
+
+
+def update(entity_dict: dict, email: str) -> Entity:
+    db: Session = next(get_db())
+    try:
+        entity_to_update: Entity = db.query(Entity).filter(
+            Entity._email == email).first()
+        if entity_to_update is None:
+            raise UserNotFoundError()
+
+        # 🔹 Mise à jour des champs si fournis
+        if "email" in entity_dict and entity_dict[
+                "email"] != entity_to_update.get_email():
+            entity_to_update.set_email(entity_dict["email"])
+
+        if "password" in entity_dict and entity_dict[
+                "password"] != entity_to_update.get_password():
+            entity_to_update.set_password(entity_dict["password"])
+
+        if "role" in entity_dict and entity_dict[
+                "role"] != entity_to_update.get_role():
+            entity_to_update.set_role(entity_dict["role"])
+
+        if "firstname" in entity_dict and entity_dict[
+                "firstname"] != entity_to_update.get_firstname():
+            entity_to_update.set_firstname(entity_dict["firstname"])
+
+        if "lastname" in entity_dict and entity_dict[
+                "lastname"] != entity_to_update.get_lastname():
+            entity_to_update.set_lastname(entity_dict["lastname"])
+
+        if "birth_at" in entity_dict and entity_dict[
+                "birth_at"] != entity_to_update.get_birth_at():
+            entity_to_update.set_birth_at(entity_dict["birth_at"])
+
+        # ✅ Appliquer les modifications
+        entity_update_dict: dict = entity_to_update.to_dict()
+        db.add(entity_to_update)
+        db.commit()
+        db.refresh(entity_to_update
+                   )  # Rafraîchir pour récupérer les nouvelles valeurs
+        del entity_update_dict["password"]
+        return entity_update_dict
+
     except UserNotFoundError:
         raise
     finally:
